👌 extract rumInternalContext spec
